MAJOR Ghost Change
Realized instead of needing a ghost image for each direction I only needed the eyes for each direction and a body to place it on. this greately reduced the amount of files I need and allows me to make a more modular ghost. Now when it is eaten I can display just the eyes and render the body once they've returned